+++ frontend/src/App.tsx
@@ -2,18 +2,18 @@ import React, { useEffect, useState } from "react";
 import "./App.css";
 import { Toaster } from "react-hot-toast";
 import { useDisclosure } from "@nextui-org/react";
-import CogTooth from "./assets/cog-tooth";
-import ChatInterface from "./components/ChatInterface";
-import Errors from "./components/Errors";
-import LoadMessageModal from "./components/LoadMessageModal";
-import { Container, Orientation } from "./components/Resizable";
-import Terminal from "./components/Terminal";
-import Workspace from "./components/Workspace";
-import { fetchMsgTotal } from "./services/session";
-import { initializeAgent } from "./services/settingsService";
-import Socket from "./services/socket";
-import { ResFetchMsgTotal } from "./types/ResponseType";
-import SettingsModal from "./components/settings/SettingsModal";
+import CogTooth from "src/assets/cog-tooth";
+import ChatInterface from "src/components/ChatInterface";
+import Errors from "src/components/Errors";
+import LoadMessageModal from "src/components/LoadMessageModal";
+import { Container, Orientation } from "src/components/Resizable";
+import Terminal from "src/components/Terminal";
+import Workspace from "src/components/Workspace";
+import { fetchMsgTotal } from "src/services/session";
+import { initializeAgent } from "src/services/settingsService";
+import Socket from "src/services/socket";
+import { ResFetchMsgTotal } from "src/types/ResponseType";
+import SettingsModal from "src/components/settings/SettingsModal";
 
 interface Props {
   setSettingOpen: (isOpen: boolean) => void;
