+++ frontend/src/components/Workspace.tsx
@@ -4,12 +4,12 @@ import { useTranslation } from "react-i18next";
 import { IoIosGlobe } from "react-icons/io";
 import { VscCode } from "react-icons/vsc";
 import { useSelector } from "react-redux";
-import Calendar from "../assets/calendar";
-import { I18nKey } from "../i18n/declaration";
-import { initialState as initialBrowserState } from "../state/browserSlice";
-import { initialState as initialCodeState } from "../state/codeSlice";
-import { RootState } from "../store";
-import { AllTabs, TabOption, TabType } from "../types/TabOption";
+import Calendar from "src/assets/calendar";
+import { I18nKey } from "src/i18n/declaration";
+import { initialState as initialBrowserState } from "src/state/browserSlice";
+import { initialState as initialCodeState } from "src/state/codeSlice";
+import { RootState } from "src/store";
+import { AllTabs, TabOption, TabType } from "src/types/TabOption";
 import Browser from "./Browser";
 import CodeEditor from "./CodeEditor";
 import Planner from "./Planner";
