+++ frontend/src/i18n/index.ts
@@ -2,7 +2,7 @@ import i18n from "i18next";
 import Backend from "i18next-http-backend";
 import LanguageDetector from "i18next-browser-languagedetector";
 import { initReactI18next } from "react-i18next";
-import { ArgConfigType } from "../types/ConfigType";
+import { ArgConfigType } from "src/types/ConfigType";
 
 export const AvailableLanguages = [
   { label: "English", value: "en" },
